DPAD-995: fixing pointer events to allow ad click
