+++ src/experimental/players/DesktopReskinnedArticleVideoPlayer/overlays/content/ContentPlayerOverlay.tsx
@@ -1,4 +1,5 @@
 import React, { useContext } from 'react';
+import styled from 'styled-components';
 import PlayerOverlay from 'experimental/shared/PlayerOverlay';
 import { ContentPlayerOverlayProps } from 'experimental/types';
 import ContentPlayerFullOverlay from 'experimental/players/DesktopReskinnedArticleVideoPlayer/overlays/content/ContentPlayerFullOverlay';
@@ -6,14 +7,17 @@ import ContentPlayerScrollOverlay from 'experimental/players/DesktopReskinnedArt
 import usePlaying from 'jwplayer/utils/usePlaying';
 import { PlayerContext } from 'jwplayer/players/shared/PlayerContext';
 
-const ContentPlayerOverlay: React.FC<ContentPlayerOverlayProps> = ({ isScrollPlayer }) => {
+const PlayerOverlayAllowClick = styled(PlayerOverlay)`
+	pointer-events: auto;
+`;
+
+const ContentPlayerOverlay: React.FC<ContentPlayerOverlayProps> = ({ isScrollPlayer, showOverlay }) => {
 	const { player } = useContext(PlayerContext);
 	const isPlaying = usePlaying();
-
 	return (
-		<PlayerOverlay handleOverlayClick={isPlaying ? player?.pause : player?.play}>
+		<PlayerOverlayAllowClick showOverlay={showOverlay} handleOverlayClick={isPlaying ? player?.pause : player?.play}>
 			{isScrollPlayer ? <ContentPlayerScrollOverlay /> : <ContentPlayerFullOverlay />}
-		</PlayerOverlay>
+		</PlayerOverlayAllowClick>
 	);
 };
